full functions detection, functions syntax not verified
