+++ HR_Garage/addons/Common/CfgFunctions.hpp
@@ -1,0 +1,8 @@
+class CfgFunctions {
+	class HR {
+		class init {
+			file = QPATHTOFOLDER(init);
+			class detector {};
+		};
+	};
+};
