+++ HR_Garage/addons/Common/init/fn_detector.sqf
