+++ HR_Garage/addons/Logistics/Public/fn_logistics_getVehCapacity.sqf
