+++ HR_Garage/addons/Utility/CfgFunctions.hpp
@@ -1,0 +1,12 @@
+class CfgFunctions {
+	class HR
+	{
+		class Utility
+		{
+			file = QPATHTOFOLDER(Functions);
+			class canFight {};
+			class classNameToModel {};
+			class localLog {};
+		};
+	};
+};
