+++ HR_Garage/addons/Utility/config.cpp
@@ -13,3 +13,5 @@ class CfgPatches {
         VERSION_CONFIG;
     };
 };
+
+#include "CfgFunctions.hpp"
